Extract await_job to sdk_extras module
